+++ packages/atomic/src/components/facets-v1/facet-search/facet-search-input.tsx
@@ -2,7 +2,7 @@ import {FunctionalComponent, h} from '@stencil/core';
 import SearchIcon from 'coveo-styleguide/resources/icons/svg/search.svg';
 import CloseIcon from 'coveo-styleguide/resources/icons/svg/close.svg';
 import {i18n} from 'i18next';
-import {createRipple} from '../../../utils/ripple';
+import {Button} from '../../common/button';
 
 interface FacetSearchInputProps {
   label: string;
@@ -41,18 +41,18 @@ export const FacetSearchInput: FunctionalComponent<FacetSearchInputProps> = (
           <atomic-icon class="w-3" icon={SearchIcon}></atomic-icon>
         </div>
         {props.query !== '' && (
-          <button
+          <Button
+            style="text-transparent"
             title={clear}
             part="search-clear-button"
-            class="search-clear-button absolute border border-transparent inline-flex justify-center items-center right-0 w-9 h-full text-on-background rounded focus:text-primary hover:text-primary focus:outline-color"
-            onMouseDown={(e) => createRipple(e, {color: 'neutral-light'})}
+            class="search-clear-button absolute inline-flex justify-center items-center right-px w-9 top-px bottom-px"
             onClick={() => {
               props.onClear();
               inputRef!.focus();
             }}
           >
             <atomic-icon class="w-2.5" icon={CloseIcon}></atomic-icon>
-          </button>
+          </Button>
         )}
       </div>
     </div>